move decodeEntities to utils
